additional improvements to cluster bootstrap
* nodes now can figure out their address to join, even if random
* all tests green
* dont use ru.smslv akka-dns
  working without the external DNS library, on A records + fixes demo app
* add docs

(cherry picked from commit 17338dfd8831215e6c832f472b0fe6d1190367ca)
