+++ cluster-http/src/main/resources/reference.conf
@@ -46,7 +46,7 @@ akka.cluster.bootstrap {
 
     # Config path of discovery method to be used to locate the initial contact points.
     # It must be a fully qualified config path to the discovery's config section.
-    discovery-method = akka.discovery.dns-srv
+    discovery-method = akka.discovery.akka-dns
 
     # Amount of time for which a discovery observation must remain "stable"
     # (i.e. not change list of discovered contact-points) before a join decision can be made.
@@ -76,7 +76,7 @@ akka.cluster.bootstrap {
     # it will assume that the the observation is "stable" (i.e. will not change),
     # and a new cluster may need to be formed. In response to this the main bootstrap
     # coordinating process may decide to join itself, or keep waiting for discovering of a seed node.
-    no-seeds-stable-margin = 10 seconds
+    no-seeds-stable-margin = 5 seconds
 
     # Interval at which contact points should be polled
     # the effective interval used is this value plus the same value multiplied by the jitter value
@@ -93,8 +93,8 @@ akka.cluster.bootstrap {
 
 akka.discovery {
 
-  dns-srv {
-    class = akka.discovery.DnsSrvServiceDiscovery
+  akka-dns {
+    class = akka.discovery.AkkaDnsServiceDiscovery
 
     resolve-interval = 1 second
