+++ cluster-http/src/test/java/akka/cluster/http/management/ClusterHttpManagementJavaCompileTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2016 Lightbend Inc. <http://www.lightbend.com>
+ * Copyright (C) 2017 Lightbend Inc. <http://www.lightbend.com>
  */
 package akka.cluster.http.management;
