+++ joining-demo/build.sbt
@@ -4,9 +4,12 @@ enablePlugins(JavaServerAppPackaging)
 
 dockerEntrypoint ++= Seq(
   """-Dakka.remote.netty.tcp.hostname="$(eval "echo $AKKA_REMOTING_BIND_HOST")"""",
-  "-Dakka.io.dns.resolver=async-dns",
-  "-Dakka.io.dns.async-dns.resolve-srv=true",
-  "-Dakka.io.dns.async-dns.resolv-conf=on"
+  """-Dakka.cluster.http.management.hostname="$(eval "echo $AKKA_REMOTING_BIND_HOST")""""
+// THIS WOULD ONLY BE NEEDED IF WE USED THE ASYNC-DNS CODE
+//,
+//  "-Dakka.io.dns.resolver=async-dns",
+//  "-Dakka.io.dns.async-dns.resolve-srv=true",
+//  "-Dakka.io.dns.async-dns.resolv-conf=on"
 )
 
 dockerCommands :=
@@ -15,7 +18,7 @@ dockerCommands :=
     case v => Seq(v)
   }
 
-version := "1.3.3.7" // FIXME
+version := "1.3.3.7" // we hard-code the version here, it could be anything really
 
 dockerUsername := Some("ktoso")
 
@@ -23,7 +26,9 @@ dockerUsername := Some("ktoso")
 dockerCommands += Cmd("USER", "root")
 
 // use ++= to merge a sequence with an existing sequence
-dockerCommands ++= Seq(
-  ExecCmd("RUN", "apt-get", "update"),
-  ExecCmd("RUN", "apt-get", "install", "-y", "dnsutils")
-)
+//
+// ENABLE THESE IF YOU WANT TO MANUALLY DO DNSLOOKUPS IN THE CONTAINER (FOR DEBUGGING)
+//dockerCommands ++= Seq(
+//  ExecCmd("RUN", "apt-get", "update"),
+//  ExecCmd("RUN", "apt-get", "install", "-y", "dnsutils")
+//)
