+++ joining-demo/src/main/resources/application.conf
@@ -1,0 +1,40 @@
+akka {
+  actor {
+    provider = "cluster"
+  }
+  remote {
+    enabled-transports = ["akka.remote.netty.tcp"]
+  }
+
+//  # <<<< ru.smslv.akka-dns library >>>>
+//  # need to force that library to be used:
+//  akka.io.dns.resolver = async-dns
+//  # we do want to resovle SRV records if available
+//  akka.io.dns.async-dns.resolve-srv = on
+//  # do use resolv conf -- no reason not to do so really
+//  akka.io.dns.async-dns.resolv-conf = on
+//  # <<<< end of smslv.akka-dns library >>>>
+
+}
+
+
+akka.cluster.http.management {
+  port = 19999
+}
+
+akka.cluster.bootstrap {
+
+  contact-point-discovery {
+    service-name = "appka-service"
+    service-namespace = "default.svc.cluster.local"
+  }
+
+  contact-point {
+    # currently this port HAS TO be the same as the `akka.cluster.http.management.port`
+    # it would not have to be once we implement the SRV record watching, since then we could potentially
+    # get the ports from the DNS records.
+    fallback-port = 19999
+
+    no-seeds-stable-margin = 3 seconds
+  }
+}
